CategoryManagerController에서 가게 조회의 책임 제거
