+++ src/main/java/com/delivery/shop/category/CategoryManagerController.java
@@ -3,12 +3,17 @@ package com.delivery.shop.category;
 import static com.delivery.shop.category.CreateCategoryParam.*;
 
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.delivery.shop.category.EditCategoryParam.EditCategoryRequest;
 
+@RequestMapping("/categories")
 public interface CategoryManagerController {
     
     /**
@@ -16,16 +21,8 @@ public interface CategoryManagerController {
      *
      * @return 카테고리 정보가 담긴 list
      */
+    @GetMapping
     ResponseEntity<GetAllCategoriesResponse> getAllCategories();
-    
-    /**
-     * url에 주어진 카테고리 id로 카테고리에 해당하고 영업중인 가게들을 조회 (휴무인 가게는 제외)
-     *
-     * @param id 카테고리 id
-     * @return 카테고리에 해당되는 가게의 정보 list
-     * @see mybatis/mapper/CategoryMapper.xml $findShopsByCategoryId
-     */
-    ResponseEntity<GetShopsByCategoryResponse> getShopsByCategory(@PathVariable("id") Long id);
     
     /**
      * 카테고리 추가
@@ -33,21 +30,23 @@ public interface CategoryManagerController {
      * @param dto 추가할 카테고리 정보
      * @return 성공시 응답 200 OK
      */
+    @PostMapping
     ResponseEntity<?> createCategory(@RequestBody CreateCategoryRequest dto);
     
     /**
      * 카테고리 삭제
      *
-     * @param id
+     * @param id 삭제할 카테고리의 id
      * @return 성공시 응답 204 NO_CONTENT
      */
+    @DeleteMapping("/{id}")
     ResponseEntity<?> deleteCategory(@PathVariable("id") Long id);
     
     /**
      * 카테고리 수정
      *
-     * @param id
-     * @param request
+     * @param id 수정할 카테고리의 id
+     * @param request 카테고리 수정 정보
      * @return 성공시 응답 200 OK
      */
     @PatchMapping("/{id}")
