+++ src/main/java/com/delivery/shop/category/CategoryManagerService.java
@@ -2,19 +2,33 @@ package com.delivery.shop.category;
 
 import java.util.List;
 
-import com.delivery.shop.category.CreateCategoryParam.CreateCategoryRequest;
-import com.delivery.shop.shop.Shop;
-
 public interface CategoryManagerService {
     
+    /**
+     * 카테고리 조회
+     *
+     * @return Category 객체 리스트
+     */
     List<Category> getAllCategories();
     
-    List<Shop> getShopsByCategory(SearchShopByCategoryRequest param);
+    /**
+     * 카테고리 추가
+     *
+     * @param param 추가될 카테고리 정보
+     */
+    void addCategory(CreateCategoryParam param);
     
-    void addCategory(CreateCategoryRequest param);
-    
+    /**
+     * 카테고리 삭제
+     *
+     * @param id 삭제될 카테고리의 id
+     */
     void deleteCategory(Long id);
     
+    /**
+     * 카테고리 정보 변경
+     * @param param 카테고리 변경 정보
+     */
     void editCategory(EditCategoryParam param);
     
 }
