+++ src/main/java/com/delivery/shop/category/CreateCategoryParam.java
@@ -26,12 +26,8 @@ public class CreateCategoryParam {
         public CreateCategoryRequest() {
         }
     
-        public CreateCategoryRequest(String name) {
-            this.name = name;
-        }
-    
-        public CreateCategoryRequest toParam() {
-            return new CreateCategoryRequest(name);
+        public CreateCategoryParam toParam() {
+            return new CreateCategoryParam(name);
         }
     
         public String getName() {
