+++ src/main/resources/db/migration/V202106071728__init.sql
@@ -36,5 +36,3 @@ CREATE TABLE MENU_GROUP
     PRIMARY KEY (ID),
     FOREIGN KEY (RESTAURANT_ID) REFERENCES SHOP(ID)
 );
-
-INSERT INTO SHOP (NAME) VALUES ('BBQ 효자동점');
