+++ src/main/resources/db/migration/V202106091544__insert_shop.sql
@@ -5,6 +5,7 @@ INSERT INTO CATEGORY (ID, NAME) VALUES (4, 'CHICKEN');
 INSERT INTO CATEGORY (ID, NAME) VALUES (5, 'PIZZA');
 INSERT INTO CATEGORY (ID, NAME) VALUES (6, 'FAST_FOOD');
 
+INSERT INTO SHOP (NAME, PHONE) VALUES ('BBQ 효자동점', '0212349999'); -- 1
 INSERT INTO SHOP (NAME, PHONE) VALUES ('도깨비한방통닭', '0212341234'); -- 2
 INSERT INTO SHOP (NAME, PHONE) VALUES ('교촌치킨 광화문점', '0238382929'); -- 3
 INSERT INTO SHOP (NAME, PHONE) VALUES ('지노스피자', '0243214321'); -- 4
